requirements.txt
pandas
numpy
yfinance
pyarrow
fastapi
uvicorn[standard]
streamlit
duckdb
qdrant-client
python-dotenv
